+++ tests/unit/test_save_checks_in_local_file.py
@@ -7,8 +7,8 @@ TEST_CHECKS = [
 ]
 
 
-def test_save_checks_to_local_file(temp_yml_file):
-    file = temp_yml_file
+def test_save_checks_to_local_file(make_local_check_file_as_yml):
+    file = make_local_check_file_as_yml
     DQEngine.save_checks_in_local_file(TEST_CHECKS, file)
     checks = DQEngine.load_checks_from_local_file(file)
     assert checks == TEST_CHECKS, "The loaded checks do not match the expected checks."
